testing

+++ Server/Dbarone.Net.CommentarioServer/CommentarioServer/DocumentGenerators/DocumentGenerator.cs
@@ -17,16 +17,6 @@ public abstract class DocumentGenerator
     #region Properties
 
     /// <summary>
-    /// The document output format.
-    /// </summary>
-    public OutputType OutputType { get; set; }
-
-    /// <summary>
-    /// The xml comments path.
-    /// </summary>
-    public string XmlCommentsPath { get; set; } = default!;
-
-    /// <summary>
     /// The assembly file path.
     /// </summary>
     public string AssemblyPath { get; set; } = default!;
@@ -37,9 +27,34 @@ public abstract class DocumentGenerator
     public string OutputPath { get; set; } = default!;
 
     /// <summary>
+    /// The document output format.
+    /// </summary>
+    public OutputType OutputType { get; set; }
+
+    /// <summary>
+    /// Set to true to allow the output documentation file to be overwritten. Otherwise an exception will be throw if the file exists.
+    /// </summary>
+    public bool OverwriteOutput { get; set; } = false;
+
+    /// <summary>
+    /// The xml comments path.
+    /// </summary>
+    public string? XmlCommentsPath { get; set; } = default!;
+
+    /// <summary>
     /// Optional path to an assembly readme. The contents are included at the top of the documentation file. 
     /// </summary>
-    public string ReadMePath { get; set; } = default!;
+    public string? ReadMePath { get; set; } = default!;
+
+    /// <summary>
+    /// Optional path to a styles file. For html output, the file must be a valid css file.
+    /// </summary>
+    public string? StylesPath { get; set; } = default!;
+
+    /// <summary>
+    /// The styles content.
+    /// </summary>
+    protected string Styles { init; get; } = default!;
 
     /// <summary>
     /// The xml comments.
@@ -56,16 +71,24 @@ public abstract class DocumentGenerator
     /// <param name="assemblyPath">The path to the assembly file.</param>
     /// <param name="outputPath">The path to the output documentation file.</param>
     /// <param name="outputType">The documentation format type.</param>
+    /// <param name="overwriteOutput">Set to true to allow existing output documentation file to be overwritten. Otherwise an exception will be thrown if the output file exists.
     /// <param name="xmlCommentsPath">The path to an xml comments file. This content will be merged with the reflected assembly information.</param>
     /// <param name="readMePath">The path to an optional readme file. This file should contain html content. If a file is specified, it will be included at the top of the documentation file.</param>
+    /// <param name="stylesPath">The path to an optional styles file.</param>
     /// <returns>Returns a <see cref="DocumentGenerator"/> instance.</returns>
     /// <exception cref="Exception">Throws an exception if invalid parameters are provided.</exception>
-    public static DocumentGenerator Create(string assemblyPath, string outputPath, OutputType? outputType = OutputType.html, string? xmlCommentsPath = null, string? readMePath = null)
+    public static DocumentGenerator Create(string assemblyPath, string outputPath, OutputType outputType = OutputType.html, bool overwriteOutput = false, string? xmlCommentsPath = null, string? readMePath = null, string? stylesPath = null)
     {
         switch (outputType)
         {
             case OutputType.html:
-                return new HtmlDocumentGenerator(xmlCommentsPath, assemblyPath, readMePath, outputPath);
+                return new HtmlDocumentGenerator(
+                    assemblyPath,
+                    outputPath,
+                    overwriteOutput,
+                    xmlCommentsPath,
+                    readMePath,
+                    stylesPath);
             default:
                 throw new Exception($"Output type {outputType.ToString()} not supported.");
         }
@@ -74,14 +97,32 @@ public abstract class DocumentGenerator
     /// <summary>
     /// Creates a new DocumentGenerator instance.
     /// </summary>
-    internal DocumentGenerator(string xmlCommentsPath, string assemblyPath, string readMePath, string outputPath)
+    internal DocumentGenerator(
+        string assemblyPath,
+        string outputPath,
+        bool overwriteOutput,
+        string? xmlCommentsPath,
+        string? readMePath,
+        string? stylesPath)
     {
-        ReadMePath = readMePath;
-        XmlCommentsPath = xmlCommentsPath;
         AssemblyPath = assemblyPath;
         OutputPath = outputPath;
+        OverwriteOutput = overwriteOutput;
+        XmlCommentsPath = xmlCommentsPath;
+        ReadMePath = readMePath;
+        StylesPath = stylesPath;
 
-        // Get documentation
+        // styles
+        if (this.StylesPath != null && File.Exists(this.StylesPath))
+        {
+            this.Styles = File.ReadAllText(this.StylesPath);
+        }
+        else
+        {
+            this.Styles = this.DefaultStyles;
+        }
+
+        // Get xml comments document.
         Comments = new XmlCommentsReader(this.XmlCommentsPath).Document;
     }
 
@@ -99,6 +140,11 @@ public abstract class DocumentGenerator
         if (!string.IsNullOrEmpty(this.ReadMePath) && !File.Exists(this.ReadMePath))
         {
             throw new Exception($"ReadMe file [{this.ReadMePath}] not found!");
+        }
+
+        if (!string.IsNullOrEmpty(this.StylesPath) && !File.Exists(this.StylesPath))
+        {
+            throw new Exception($"Styles file [{this.ReadMePath}] not found!");
         }
     }
 
@@ -312,28 +358,36 @@ public abstract class DocumentGenerator
 
     protected ConstructorInfo[] GetConstructors(Type type)
     {
-        return type.GetConstructors(BindingFlags.DeclaredOnly | BindingFlags.Public | BindingFlags.NonPublic | BindingFlags.Instance | BindingFlags.Static);
+        return type.GetConstructors(BindingFlags.DeclaredOnly | BindingFlags.Public | BindingFlags.NonPublic | BindingFlags.Instance | BindingFlags.Static)
+        .Where(c => !IsPrivate(c)).ToArray();
     }
 
     protected MethodInfo[] GetMethods(Type type)
     {
         // ignore getter/setter methods.
-        return type.GetMethods(BindingFlags.DeclaredOnly | BindingFlags.Public | BindingFlags.NonPublic | BindingFlags.Instance | BindingFlags.Static).Where(m => !m.IsSpecialName).ToArray();
+        return type.GetMethods(BindingFlags.DeclaredOnly | BindingFlags.Public | BindingFlags.NonPublic | BindingFlags.Instance | BindingFlags.Static)
+            .Where(m => !m.IsSpecialName)
+            .Where(m => !IsPrivate(m)).ToArray()
+            .ToArray();
     }
 
     protected FieldInfo[] GetFields(Type type)
     {
-        return type.GetFields(BindingFlags.DeclaredOnly | BindingFlags.Public | BindingFlags.NonPublic | BindingFlags.Instance | BindingFlags.Static);
+        return type.GetFields(BindingFlags.DeclaredOnly | BindingFlags.Public | BindingFlags.NonPublic | BindingFlags.Instance | BindingFlags.Static)
+                .Where(f => !IsPrivate(f)).ToArray();
     }
 
     protected PropertyInfo[] GetProperties(Type type)
     {
-        return type.GetProperties(BindingFlags.DeclaredOnly | BindingFlags.Public | BindingFlags.NonPublic | BindingFlags.Instance | BindingFlags.Static);
+        return type.GetProperties(BindingFlags.DeclaredOnly | BindingFlags.Public | BindingFlags.NonPublic | BindingFlags.Instance | BindingFlags.Static)
+                .Where(p => !IsPrivate(p)).ToArray();
+
     }
 
     protected EventInfo[] GetEvents(Type type)
     {
-        return type.GetEvents(BindingFlags.DeclaredOnly | BindingFlags.Public | BindingFlags.NonPublic | BindingFlags.Instance | BindingFlags.Static);
+        return type.GetEvents(BindingFlags.DeclaredOnly | BindingFlags.Public | BindingFlags.NonPublic | BindingFlags.Instance | BindingFlags.Static)
+                .Where(e => !IsPrivate(e)).ToArray();
     }
 
     protected string GetReadMe()
@@ -369,6 +423,9 @@ public abstract class DocumentGenerator
         ValidateParameters();
 
         // Save file
+        if (this.OverwriteOutput==false && File.Exists(this.OutputPath)) {
+            throw new Exception("Output file already exists and [OverwriteOutput] parameter set to false.");
+        }
         File.WriteAllText(this.OutputPath, RenderDocument());
     }
 
@@ -377,306 +434,25 @@ public abstract class DocumentGenerator
     protected abstract string RenderType(Type type);
     protected abstract string RenderTypeTOCSection(Type type, string header, MemberInfo[] members);
     protected abstract string RenderTypeMember(MemberInfo member);
-
     protected abstract string RenderItems(object[] items);
-
     protected abstract string RenderExample(ExampleNode node);
-
     protected abstract string RenderCode(CodeNode node);
-
     protected abstract string RenderC(CNode node);
-
     protected abstract string RenderPara(ParaNode node);
-
     protected abstract string RenderSee(SeeNode node);
-
     protected abstract string RenderReturns(ReturnsNode node);
-
     protected abstract string RenderExceptions(ExceptionNode[] nodes);
     protected abstract string RenderException(ExceptionNode node);
     protected abstract string RenderParam(ParamNode? node);
-
     protected abstract string RenderList(ListNode node);
+    protected abstract string DefaultStyles { get; }
 
     /// <summary>
     /// Renders the type's generic arguments.
     /// </summary>
     /// <param name="type">The type to render.</param>
     /// <returns>A string representing the generic arguments.</returns>
-
     protected abstract string RenderTypeGenericArguments(Type type);
-
-    protected string GetCSSStyles()
-    {
-        var css = @"
-<style type=""text/css"">
-
-    /* https://goodpalette.io */
-    :root {
-
-        /* Primary: Blue Blue */
-        --primary-100: #F3F2FF;
-        --primary-200: #C0C0FE;
-        --primary-300: #8B91F9;
-        --primary-400: #5565E9;
-        --primary-500: #233FCC;
-        --primary-600: #0C2EA3;
-        --primary-700: #032479;
-        --primary-800: #001B50;
-        --primary-900: #000F26;
-
-        /* Accent: Jaffa */
-        --accent-100: #FFF9F2;
-        --accent-200: #FFE4C8;
-        --accent-300: #FBC89D;
-        --accent-400: #F3A470;
-        --accent-500: #E37944;
-        --accent-600: #B44921;
-        --accent-700: #852A11;
-        --accent-800: #551509;
-        --accent-900: #260704;
-
-        /* Neutral */
-        --neutral-100: #FAFAFC;
-        --neutral-200: #EAEAEF;
-        --neutral-300: #DADAE2;
-        --neutral-400: #CACBD4;
-        --neutral-500: #BBBDC7;
-        --neutral-600: #94969F;
-        --neutral-700: #6D7077;
-        --neutral-800: #474A4E;
-        --neutral-900: #222426;
-
-        --white: #fff;
-    }
-
-    /* -----------------------------------------------
-    Base Styles
-    -------------------------------------------------- */
-
-    body {
-        font-family: ""Helvetica Neue"", Helvetica, Arial, sans-serif;
-        color: var(--neutral-900);
-        overflow-y: scroll;
-        font-size: 0.9em;
-    }
-
-    div.toc {
-        background-color: var(--primary-200);
-        border: 1px solid var(--primary-500);
-        border-radius: 4px;
-        margin: 4px 0px;
-        padding: 0px;
-    }
-
-    div.toc-inner {
-        margin: 4px;
-    }
-
-    div.type {
-        background-color: var(--accent-200);
-        border: 1px solid var(--accent-500);
-        border-radius: 4px;
-        margin: 4px 0px;
-        padding: 0px;
-    }
-
-    div.type-inner {
-        margin: 4px;
-    }
-
-    div.member {
-        background-color: var(--neutral-200);
-        border: 1px solid var(--neutral-500);
-        border-radius: 4px;
-        margin: 4px 0px;
-        padding: 4px 4px;
-    }
-
-    div.member-inner {
-        margin: 4px;
-    }
-
-    /* ------------------------------------
-    Typography
-    --------------------------------------- */
-
-    h1, h2, h3, h4, h5, h6 {
-        font-weight: 500;
-    }
-
-    h1 {
-        font-size: 2.0em;
-        display: block;
-        background: var(--primary-300);
-        border: 1px solid var(--primary-700);
-        border-left: 2em solid var(--primary-700);
-        border-right: 2em solid var(--primary-700);
-        padding: 24px 6px;
-        border-radius: 4px;
-    }
-
-    h2 {
-        font-size: 1.8em;
-    }
-
-    h3 {
-        font-size: 1.6em;
-        margin: 1em 0em 0em 0em;
-    }
-
-    div.toc h2  {
-        border-left: 2.0em solid var(--primary-500);
-        display: block;
-        padding: 12px 2px 12px 12px;
-        background-color: var(--primary-300);
-        margin: 0px;
-    }
-
-    div.toc h3 {
-        color: var(--primary-500);
-    }
-
-    div.type h2  {
-        border-left: 2.0em solid var(--accent-500);
-        display: block;
-        padding: 12px 2px 12px 12px;
-        background-color: var(--accent-300);
-        margin: 0px;
-    }
-
-    div.type h3 {
-        color: var(--accent-500);
-    }
-
-    div.member h2  {
-        border-left: 2.0em solid var(--primary-500);
-        display: block;
-        padding: 12px 2px 12px 12px;
-        background-color: var(--primary-300);
-        margin: 0px;
-    }
-
-    div.member h3 {
-        color: var(--primary-500);
-    }
-
-    /* -----------------------------------
-    Links
-    -------------------------------------- */
-
-    a {
-        color: var(--accent-700);
-        text-decoration: none;
-    }
-
-    a:hover {
-        color: var(--accent-700);
-        text-decoration: underline;
-    }
-
-    /* ---------------------------------------------
-    Lists
-    ------------------------------------------------ */
-
-    ol, ul {
-        padding: 0px;
-        margin: 0px;
-        margin-bottom: 0.5em;
-    }
-
-    ul {
-        list-style: disc inside;
-    }
-
-    ol {
-        list-style: decimal inside;
-    }
-
-    /* ---------------------------------------------
-    Code / Pre
-    ------------------------------------------------ */
-
-    pre code {
-        background-color: var(--white);
-        border: 1px solid var(--neutral-700);
-        border-left: 6px solid var(--accent-700);
-        color: var(--neutral-900);
-        page-break-inside: avoid;
-        font: ""Courier New"", Courier, Monospace;
-        max-width: 100%;
-        padding: 2em 1em;
-        display: inline-block;
-        word-wrap: break-word;
-        overflow: auto;
-        overflow-x: auto;
-        white-space: pre-wrap;
-        font-weight: 500;
-    }
-
-    code {
-        font-weight: bold;
-        background-color: var(--neutral-500);
-        padding: 2px;
-    }
-
-    /* ---------------------------------------------
-    Tables
-    ------------------------------------------------ */
-
-    div.table {
-        border: 1px solid var(--primary-700);
-        display: inline-block;
-        border-radius: 4px;
-        padding: 0px;
-        margin: 0px;
-        overflow: hidden
-    }
-
-    table {
-        border-collapse: separate;
-        border-spacing: 0;
-        padding: 0px;
-    }
-
-    th,
-    td {
-        padding: 6px 24px;
-        text-align: left;
-        margin: 0px;
-    }
-
-    th {
-        background-color: var(--primary-700);
-        color: var(--neutral-300);
-        font-weight: 500;
-    }
-
-    tr:nth-child(odd) {
-        background-color: var(--neutral-100);  
-    }
-
-    /* ---------------------------------------------
-    Spacing
-    ------------------------------------------------ */
-
-    blockquote,
-    dl,
-    figure,
-    table,
-    p,
-    ul,
-    ol,
-    {
-        margin-top: 0em;
-        margin-bottom: .5em;
-    }
-
-</style>
-        ";
-
-        return css;
-    }
 
     /// <summary>
     /// Returns a hyperlink for a type.
@@ -686,7 +462,7 @@ public abstract class DocumentGenerator
     protected string GetLinkForType(Type type)
     {
         var types = this.GetTypes();
-        if (types.Select(t=>t.ToCommentId()).Contains(type.ToCommentId()))
+        if (types.Select(t => t.ToCommentId()).Contains(type.ToCommentId()))
         {
             return @$"<a href=""#{type.ToCommentId()}"">{type.Name}</a>";
         }
